Add files via upload
This is my final year BCA project ,Restaurant Table Booking System, the project is developed using HTML, CSS & JS as the front-end , PHP as the back-end , MYSQL as the database and to run the project i used a local host as the XAMPP server. This project is developed where the customers can book the table from any restaurant of any state where they wish to, they can also view the menu and also pre-order the food simultaneously.